src/ui/requirements.txt
agent-framework
openai
asyncio
python-dotenv
azure-search-documents
fastapi
pandas
uvicorn
streamlit
requests
azure-identity